Add files via upload
I first tried running xgboost in R but since xgboost requires dummification of categorical variables and the categorical variables in the dataset have over 1,000 levels, the running efficiency is low.
I used Python afterwards with xgboost library. It seems that it is more flexible in terms of customizing the evaluation metric.